requirements.txt
tensorflow
google-cloud-aiplatform